+++ src/main/java/gift/controller/WishlistController.java
@@ -3,8 +3,7 @@ package gift.controller;
 import gift.service.WishlistService;
 import gift.model.Product;
 import gift.model.User;
-import gift.util.LoginUser;
-import gift.controller.WishlistRequest;
+import gift.annotation.LoginUser;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.*;
