+++ src/main/java/gift/model/Wishlist.java
@@ -1,4 +1,5 @@
 package gift.model;
 
 public record Wishlist(Long id, String userEmail, Long productId) {
+
 }
